Followup: Move OptionalEmitter into dedicated file
Gettings things back into a state that resembles upstream SpiderMonkey somewhat
more closely.
