+++ js/src/frontend/BytecodeEmitter.cpp
@@ -37,6 +37,7 @@
 #include "frontend/ForOfLoopControl.h"
 #include "frontend/IfEmitter.h"
 #include "frontend/NameOpEmitter.h"
+#include "frontend/OptionalEmitter.h"
 #include "frontend/Parser.h"
 #include "frontend/PropOpEmitter.h"
 #include "frontend/TDZCheckCache.h"
@@ -74,144 +75,6 @@ ParseNodeRequiresSpecialLineNumberNotes(ParseNode* pn)
 {
     return pn->getKind() == ParseNodeKind::While || pn->getKind() == ParseNodeKind::For;
 }
-
-// OOOEEE start
-
-class MOZ_RAII BytecodeEmitter::OptionalEmitter {
-  public:
-    enum class Kind {
-        // Requires two values on the stack
-        Reference,
-        // Requires one value on the stack
-        Other
-    };
-
-  private:
-    BytecodeEmitter* bce_;
-
-    TDZCheckCache tdzCache_;
-
-    // jumpTarget for the fake label over the optional chaining code
-    JumpList top_;
-
-    // BreakableControl target for the break from inside the optional chaining code
-    BreakableControl breakInfo_;
-
-    int32_t initialDepth_;
-
-    // JSOp is the op code to be emitted, Kind is if we are dealing with a
-    // reference (in which case we need two elements on the stack) or other value
-    // (which needs one element on the stack)
-    JSOp op_;
-    Kind kind_;
-
-  public:
-    OptionalEmitter(BytecodeEmitter* bce, int32_t initialDepth, JSOp op = JSOP_UNDEFINED, Kind kind = Kind::Other)
-        : bce_(bce), tdzCache_(bce),
-          breakInfo_(bce, StatementKind::Label),
-          initialDepth_(initialDepth), op_(op), kind_(kind) {}
-
-    bool emitOptionalJumpLabel() {
-
-        return bce_->emitJump(JSOP_LABEL, &top_);
-    }
-
-    bool emitJumpShortCircuit() {
-        MOZ_ASSERT(initialDepth_ + 1 == bce_->stackDepth);
-        InternalIfEmitter ifEmitter(bce_);
-        if (!bce_->emitPushNotUndefinedOrNull())
-            return false;
-
-        if (!bce_->emit1(JSOP_NOT))
-            return false;
-
-        if (!ifEmitter.emitThen())
-            return false;
-
-        // Perform ShortCircuiting code and break
-        if (!bce_->emit1(JSOP_POP))
-            return false;
-
-        if (!bce_->emit1(op_))
-            return false;
-
-        if (kind_ == Kind::Reference) {
-            if (!bce_->emit1(op_))
-                return false;
-        }
-
-        if (!bce_->emitGoto(&breakInfo_, &breakInfo_.breaks, SRC_BREAK2LABEL))
-            return false;
-
-        if (!ifEmitter.emitEnd()) {
-            return false;
-        }
-        return true;
-    }
-
-    bool emitJumpShortCircuitForCall() {
-        int32_t depth = bce_->stackDepth;
-        MOZ_ASSERT(initialDepth_ + 2 == depth);
-        if (!bce_->emit1(JSOP_SWAP))
-            return false;
-
-        InternalIfEmitter ifEmitter(bce_);
-        if (!bce_->emitPushNotUndefinedOrNull())
-            return false;
-
-        if (!bce_->emit1(JSOP_NOT))
-            return false;
-
-        if (!ifEmitter.emitThen())
-            return false;
-
-        // Perform ShortCircuiting code for Call and break
-        if (!bce_->emit1(JSOP_POP))
-            return false;
-
-        if (!bce_->emit1(JSOP_POP))
-            return false;
-
-        if (!bce_->emit1(op_))
-            return false;
-
-        if (kind_ == Kind::Reference) {
-            if (!bce_->emit1(op_))
-                return false;
-        }
-
-        if (!bce_->emitGoto(&breakInfo_, &breakInfo_.breaks, SRC_BREAK2LABEL))
-            return false;
-
-        if (!ifEmitter.emitEnd())
-            return false;
-
-        bce_->stackDepth = depth;
-
-        if (!bce_->emit1(JSOP_SWAP))
-            return false;
-        return true;
-    }
-
-    bool emitOptionalJumpTarget() {
-
-        // Patch the JSOP_LABEL offset.
-        JumpTarget brk{ bce_->lastNonJumpTargetOffset() };
-        bce_->patchJumpsToTarget(top_, brk);
-
-        // Patch the emitGoto() offset.
-        if (!breakInfo_.patchBreaks(bce_))
-            return false;
-
-        // XXX: Commented out due to workaround for missing JSOP_GOTO functionality
-        /*// reset stack depth to the depth when we jumped
-        bce_->stackDepth = initialDepth_ + 1;*/
-
-        return true;
-    }
-};
-
-// OOOEEE end
 
 BytecodeEmitter::BytecodeEmitter(BytecodeEmitter* parent,
                                  const EitherParser<FullParseHandler>& parser, SharedContext* sc,
